requirements.txt
absl-py==2.3.1
ale-py==0.10.1
alembic==1.14.1
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.7
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.1.1
cryptography==46.0.7
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
google-auth==2.49.2
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.70.0
gym==0.21.0
gym-notices==0.1.0
gym-retro==0.8.0
gymnasium==1.0.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.3.1
opencv-python==4.13.0.92
optuna==4.5.0
packaging==26.0
pandas==2.0.3
pillow==10.4.0
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pygame==2.6.1
pyglet==1.5.31
Pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.3.4
six==1.17.0
SQLAlchemy==2.0.49
stable_baselines3==2.4.1
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.67.3
typing_extensions==4.13.2
tzdata==2026.1
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
